server/src/python/requirements.txt
youtube-transcript-api==0.5.0
openai==1.14.1
python-dotenv==1.0.1
requests==2.31.0 